fix bug-nghiep-2
