+++ app/Http/Requests/role/StoreRoleRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\role;
 
 use Illuminate\Foundation\Http\FormRequest;
