+++ app/Http/Requests/user/StoreUserRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\user;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class UpdateUserRequest extends FormRequest
+class StoreUserRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -21,23 +21,22 @@ class UpdateUserRequest extends FormRequest
      */
     public function rules(): array
     {
-        $userID = $this->route('user');
         return [
             'name'     => 'required|max:255',
-            'image'    => 'nullable|image|mimes:jpg,jpeg,png,gif|max:2048',
-            'email'    => 'required|max:255|email|unique:users,email,' . $userID,
-            'password' => 'required|min:6',
-            'phone'    => 'required|max:15|min:10|unique:users,phone,' . $userID,
+            'image'    => 'required|image|mimes:jpg,jpeg,png,gif|max:2048',
+            'email'    => 'required|max:255|email|unique:users,email,',
+            'password' => 'required',
+            'phone'    => 'required|max:15|min:10|unique:users,phone,',
             'role_id'  => 'nullable',
-            'active'   => 'nullable',
         ];
     }
 
     public function messages()
     {
         return [
-            'name.required'     => 'Tên không được để trống',
+           'name.required'      => 'Tên không được để trống',
             'name.max'          => 'Tên không được dài quá 255 ký tự',
+            'image.required'    => 'Tên không được để trống',
             'image.image'       => 'File phải là định dạng hình ảnh',
             'image.mimes'       => 'Ảnh phải có định dạng jpg, jpeg, png hoặc gif',
             'image.max'         => 'Ảnh không được vượt quá 2MB',
@@ -49,8 +48,8 @@ class UpdateUserRequest extends FormRequest
             'password.min'      => 'Mật khẩu phải có ít nhất 6 ký tự',
             'phone.required'    => 'Số điện thoại không được để trống',
             'phone.unique'      => 'Số điện thoại đã tồn tại',
-            'phone.max'         => 'Số điện thoại tối thiểu 10 số',
-            'phone.min'         => 'Số điện thoại tối đa 15 số',
+            'phone.min'         => 'Số điện thoại tối thiểu 10 số',
+            'phone.max'         => 'Số điện thoại tối đa 15 số',
         ];
     }
 }
