+++ app/Http/Requests/user/UpdateUserRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\user;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class StoreUserRequest extends FormRequest
+class UpdateUserRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -21,20 +21,22 @@ class StoreUserRequest extends FormRequest
      */
     public function rules(): array
     {
+        $userID = $this->route('user');
         return [
             'name'     => 'required|max:255',
             'image'    => 'nullable|image|mimes:jpg,jpeg,png,gif|max:2048',
-            'email'    => 'required|max:255|email|unique:users,email,',
-            'password' => 'required',
-            'phone'    => 'required|max:15|min:10|unique:users,phone,',
+            'email'    => 'required|max:255|email|unique:users,email,' . $userID,
+            'password' => 'required|min:6',
+            'phone'    => 'required|max:15|min:10|unique:users,phone,' . $userID,
             'role_id'  => 'nullable',
+            'active'   => 'nullable',
         ];
     }
 
     public function messages()
     {
         return [
-           'name.required'     => 'Tên không được để trống',
+            'name.required'     => 'Tên không được để trống',
             'name.max'          => 'Tên không được dài quá 255 ký tự',
             'image.image'       => 'File phải là định dạng hình ảnh',
             'image.mimes'       => 'Ảnh phải có định dạng jpg, jpeg, png hoặc gif',
