+++ resources/views/page/role/create.blade.php
@@ -13,11 +13,11 @@
                     <form action="{{ route('role.store') }}" method="POST">
                         @csrf
                         <div class="form-group">
-                            <label for="name" class="control-label mb-1">Tên vai trò *</label>
+                            <label for="name" class="control-label mb-1">Tên vai trò</label>
                             <input type="text" id="name" name="name" class="form-control"
                                 value="{{ old('name') }}">
                             @error('name')
-                                <p class="text-danger">{{ $message }}</p>
+                                <p class="text-danger badge">{{ $message }}</p>
                             @enderror
                         </div>
